Tidy eft_utils imports
Put eft_utils functions in alphabetical order. Move an unrelated function.
