+++ scaffoldmaker/utils/eftfactory_tricubichermite.py
@@ -4,7 +4,7 @@ Created on Nov 15, 2017
 
 @author: Richard Christie
 '''
-from scaffoldmaker.utils.eft_utils import *
+from scaffoldmaker.utils.eft_utils import mapEftFunction1Node1Term, remapEftLocalNodes, remapEftNodeValueLabel, scaleEftNodeValueLabels, setEftScaleFactorIds
 from scaffoldmaker.utils import interpolation as interp
 from scaffoldmaker.utils import zinc_utils
 from scaffoldmaker.utils import vector
@@ -13,6 +13,20 @@ from opencmiss.zinc.field import Field
 from opencmiss.zinc.node import Node
 from opencmiss.zinc.status import OK as ZINC_OK
 import math
+
+def derivativeSignsToExpressionTerms(valueLabels, signs):
+    '''
+    Return remap expression terms for summing derivative[i]*sign[i]
+    :param valueLabels: List of node value labels to possibly include.
+    :param signs: List of 1 (no scaling), -1 (scale by scale factor 1) or 0 (no term).
+    '''
+    expressionTerms = []
+    for i in range(len(valueLabels)):
+        if signs[i] is 1:
+            expressionTerms.append( ( valueLabels[i], [] ) )
+        elif signs[i] is -1:
+            expressionTerms.append( ( valueLabels[i], [1] ) )
+    return expressionTerms
 
 class eftfactory_tricubichermite:
     '''
